refactor: Remove SOS tab and breathing exercise
The SOS tab and the associated breathing exercise component have been removed. This feature was intended for emergency situations, but it has been decided that its functionality is not core to the CaffeWise application's primary goal of caffeine management and prediction. The related UI elements and routing have been cleaned up.